separated spec files for feature
