+++ cypress/integration/v2/blocks/feature/mobile-advanced.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { mobileAdvanced } from '.'
+
+describe( 'Feature Block ( Mobile Advanced )', registerTests( [ mobileAdvanced ] ) )
