+++ cypress/integration/v2/blocks/feature/mobile-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { mobileStyle } from '.'
+
+describe( 'Feature Block ( Mobile Style )', registerTests( [ mobileStyle ] ) )
